backend/requirements.txt
asgiref==3.4.1
Django==3.2
django-cors-headers==3.13.0
django-filter==21.1
djangorestframework==3.14.0
pytz==2022.5
sqlparse==0.4.3
typing_extensions==4.1.1
